refactor: extract cancellable file search
AI-Generated: ChatGPT (OpenAI)
